[refactor] avoid a warning
The warning "`[WARNING]: The value True (type bool) in a string field 
was converted to u'True' (type string). If this does not look like what 
you expect, quote the
entire value to ensure it does not change.`" is caused by the `upgrade: 
yes`: as this option can takes some other arguments as `dist`, `full` or 
`safe`, it have to be handle as a string. See 
https://serverfault.com/a/993211.